refactor(text-to-image): separates concern of user-facing messaging from TextToImageOutputAlert
